cpp02 compiles and passes

+++ cpp02/ex03/Point.cpp
@@ -6,17 +6,18 @@
 /*   By: bamrouch <bamrouch@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/16 14:10:00 by bamrouch          #+#    #+#             */
-/*   Updated: 2023/09/20 19:31:44 by bamrouch         ###   ########.fr       */
+/*   Updated: 2023/09/22 18:58:03 by bamrouch         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "Point.hpp"
 
-Point::Point():x((float)0.0), y((float)0.0)
+Point::Point():x(0.0f), y(0.0f)
 {};
 
 Point::Point(float const new_x, float const new_y):x(new_x), y(new_y)
-{};
+{
+};
 
 Point::Point(const Point &cpy_point):x(cpy_point.x), y(cpy_point.x)
 {};
